created like-dislike functionality on server side and model side. Still working ot front end problems with it

+++ db/models/post.js
@@ -8,10 +8,10 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
       },
     total_likes: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.INTEGER,
       },
     total_comments: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.INTEGER,
     }
   }, {});
   Post.associate = function(models) {
